fixed builds where a pure offline yarn.lock was pulled in
During this build, yarn connects to the internet to download
dependencies. However, while in online mode and processing a
yarn.lock file entry that is offline (i.e. it was resolved
from the offline cache and not the registry), yarn will not
connect to the internet to grab it.

This is a problem because projects that use this library can
and should use such offline yarn.lock files.

Our simple workaround is to preprocess the yarn.lock file and
change any offline cache entries to registry entries.

Change-Id: I28930a7d768daa14e4f37587a6370ed693950969
Signed-off-by: Greg Sheremeta <gshereme@redhat.com>