add pre-seed feature to fix chicken and egg problem
When working on a project that depends on ovirt-engine-nodejs-modules,
there is a CI chicken and egg problem when you update dependencies.
Your project gets its dependencies from ovirt-engine-nodejs-modules.
However, ovirt-engine-nodejs-modules directly references *already
committed* versions of dependent projects' package.json and yarn.lock
files. If you want to update or add dependencies, they must first
exist in ovirt-engine-nodejs-modules. But your project must pass
CI to get merged, but it can't because it can't get its new
dependencies... Chicken and egg.

Previously, the way to work around this was either:
1. submit a patch to your project containing only required package.json
   and yarn.lock changes, then rebuild ovirt-engine-nodejs-modules
   Now your actual patch can get the new dependencies
   from ovirt-engine-nodejs-modules
2. force submit your patch even though it's failing CI.
   Then rebuild ovirt-engine-nodejs-modules. Then rebuild
   your patch and it should now pass.

Neither is an ideal workaround.

Introduce a pre-seed directory to fix this problem.
When working on a new patch, add subdirectory in the pre-seed
directory, and copies of the package.json and yarn.lock your
new project or patch is using.

For example:
 - pre-seed/2018-01-14-ovirt-engine-dashboard/package.json
 - pre-seed/2018-01-14-ovirt-engine-dashboard/yarn.lock
 - pre-seed/2018-01-15-example-project/package.json
 - pre-seed/2018-01-15-example-project/yarn.lock

(The pre-seed directory can occasionally be pruned.)

After committing a patch that adds a pre-seed,
ovirt-engine-nodejs-modules will now be pre-seeded with your
new package.json + yarn.lock requirements.

Change-Id: I5d6435346c87aa1dbb9b98397e2bc101c080eeb3
Signed-off-by: Greg Sheremeta <gshereme@redhat.com>