Setup embedded yarn to be executable directly
Now that ovirt-engine-yarn is obsolete, upgrading ovirt-web-ui
to use the new setup caused a problem.  Shell aliases as defined
in `setup-env.sh` are not exported.  So while the offline cache is
setup, yarn install and yarn check run ok, using yarn outside of
setup-env.sh doesn't work.

By leveraging the hashbang in yarn-*.js, the js is installed as a
normal executable.  Embedded yarn can now easily be used in place of
any other yarn without changes to the dependent project build scripts.
As long as setup-env.sh is used, which pushes the bin/ with yarn to
the PATH, nothing else will need to be done.

Change-Id: I77cf9949052c36e55c1535cea315effe9e53b93c
Signed-off-by: Scott J Dickerson <sdickers@redhat.com>