Allow merged PRs in the build but still block check
The pre-seed verification script is updated to flag warnings
and errors and report them out with different exit codes.  Exit code
of 1 is a warning on a pre-seed PR while exit code 2 is at least one
error.  Errors happen if the pre-seed is not configured correctly
or if the important files cannot be fetched from the git repo.  A
warning happens if a pre-seed PR is closed/merged.

If a build can be blocked by a PR that is now merged, then any rebuilds
will fail.  This change has the build ignore pre-seed warnings and
continue the build.

The CI check will fail on any pre-seed warnings or errors.