refactor the build to use git and github api (#22)
* refactor the build to use git and github api

= Summary

There are two main changes with this commit:
  - Use `git` to fetch project repositories to access the files
  - Refactor pre-seed handling

All project and pre-seed settings are now held in `projects.list.mjs`.

= Using `git`

Instead of using direct fetch URLs to access the `package.json` and
`yarn.lock` files for projects, use `git` to pull the project
repository.  There are a few benefits:
  - The git fetch seems to access the files faster than the wget method
  - git history is available to include in the srpm build
  - Projects and pre-seeds via pull requests can be handled the same way

To normalize the projects, the important files (`package.json` and
`yarn.lock`) are copied from the git repo to a `projects_files/`
directory.  The target directory includes the project name, the source
branch, the fetch date and the HEAD git commit hash.  This allows us
to identify exactly where the files were pulled from and when even
after installing the rpm.

= Pre-seed handling

Pre-seeds are now defined in the `projects.list.mjs` javascript file,
and the github rest api is used to (1) find the git clone url for the
project, and (2) verify the pre-seed pull request is valid as a pre-seed
sources.  It is no longer required to manually copy the `package.json`
and `yarn.lock` files to this project.  They are fetched from the pull
requests branch in the same way as the projects are fetched.

= Other changes

  - Add GITHUB_TOKEN to automation scripts in github actions so the
    API calls can be authenticated and have a much higher call rate
    limiting

  - Github action `comment-to-build-pr.yaml` updated to:

    - install the previous version of ovirt-engine-nodejs-modules
      as a cache source

    - add fail comment with a link to the action's run in case of
      action failure

  - Add `Apache-2.0` license file for the project

* Allow merged PRs in the build but still block check

The pre-seed verification script is updated to flag warnings
and errors and report them out with different exit codes.  Exit code
of 1 is a warning on a pre-seed PR while exit code 2 is at least one
error.  Errors happen if the pre-seed is not configured correctly
or if the important files cannot be fetched from the git repo.  A
warning happens if a pre-seed PR is closed/merged.

If a build can be blocked by a PR that is now merged, then any rebuilds
will fail.  This change has the build ignore pre-seed warnings and
continue the build.

The CI check will fail on any pre-seed warnings or errors.

* Update README.adoc

* In JS, throw Errors with cause when available.

* Move pre-seed definitions to `pre-seeds.list.mjs`