Fix copr make srpm build
We install ovirt-engine-nodejs-modules most recent published version
as part of the srpm build as a way to prefill the yarn fetch cache.
The dnf repo is referenced directly via the `--repofrompath` cli
option.

The repo referenced previously is the fedora 34 version.  That repo
is no longer available.  Adjusting to a new repo source that should
be more stable should fix this problem.